+++ src/main/java/com/trepicoder/adventurecraft/blockstateproperties/WoodpileFullness.java
@@ -1,4 +1,19 @@
 package com.trepicoder.adventurecraft.blockstateproperties;
 
-public enum WoodpileFullness {
+import net.minecraft.util.IStringSerializable;
+
+public enum WoodpileFullness implements IStringSerializable {
+    EMPTY("empty"),
+    SHALLOW("shallow"),
+    HALF("half"),
+    FULL("full");
+
+    private final String name;
+
+    private WoodpileFullness(String name) { this.name = name; }
+
+    @Override
+    public String getName() {
+        return this.name;
+    }
 }