+++ src/main/java/com/trepicoder/adventurecraft/registries/ACBlockStateProperties.java
@@ -1,4 +1,8 @@
 package com.trepicoder.adventurecraft.registries;
 
+import com.trepicoder.adventurecraft.blockstateproperties.WoodpileFullness;
+import net.minecraft.state.EnumProperty;
+
 public class ACBlockStateProperties {
+    public static final EnumProperty<WoodpileFullness> WOODPILE_FULLNESS = EnumProperty.create("fullness", WoodpileFullness.class);
 }